feat: 部分软件增加安装校验

+++ scripts/redis/install.sh
@@ -45,6 +45,11 @@ else
     echo "错误：耗子面板不支持该系统"
     exit 1
 fi
+if [ "$?" != "0" ]; then
+    echo -e $HR
+    echo "错误：安装依赖软件失败，请截图错误信息寻求帮助。"
+    exit 1
+fi
 
 # 准备目录
 rm -rf ${redisPath}
